remove old internal metrics lib
